EFS-1100: Rename file
