+++ src/operations/query/r4.js
@@ -15,7 +15,7 @@ const { FieldMapper } = require('./filters/fieldMapper');
 const { FilterByMissing } = require('./filters/missing');
 const { FilterByContains } = require('./filters/contains');
 const { FilterByAbove, FilterByBelow } = require('./filters/aboveAndBelow');
-const { FilterByText } = require('./filters/partialText');
+const { FilterByText } = require('./filters/text');
 const { FilterById } = require('./filters/id');
 const { MongoQuerySimplifier } = require('../../utils/mongoQuerySimplifier');
 const { FilterParameters } = require('./filters/filterParameters');
